refactor: move PublishedStreamData class
